Move frame_errors to scylla-cql-core
